refactor: model 패키지 -> domain 패키지로 rename
