+++ src/main/java/blackjack/Application.java
@@ -1,7 +1,7 @@
 package blackjack;
 
 public class Application {
-    // 동기화 테스트
+
     public static void main(String[] args) {
         BlackjackGamePlay blackjackGamePlay = new BlackjackGamePlay();
         blackjackGamePlay.run();
