+++ src/main/java/blackjack/domain/gamer/Name.java
@@ -1,4 +1,4 @@
-package blackjack.model.gamer;
+package blackjack.domain.gamer;
 
 public class Name {
