+++ src/test/java/blackjack/domain/gamer/DealerTest.java
@@ -1,8 +1,8 @@
-package blackjack.model.gamer;
+package blackjack.domain.gamer;
 
-import blackjack.model.card.Card;
-import blackjack.model.card.Pattern;
-import blackjack.model.card.Rank;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Pattern;
+import blackjack.domain.card.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
