+++ src/test/java/blackjack/domain/gamer/NameTest.java
@@ -1,4 +1,4 @@
-package blackjack.model.gamer;
+package blackjack.domain.gamer;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
